made post user extractor middleware work again

+++ part4/bloglistapp/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
 
@@ -10,16 +10,10 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   try {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)  
-    if (!decodedToken.id) {    
-      return response.status(401).json({ 
-        error: 'token invalid' 
-      })  
-    }  
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
       title: body.title,
@@ -39,8 +33,32 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  // can only delete blog if you are the owner of the blog
+  // token sent with the request must match that of the blog's creator
+  // could compare user id's instead?
+
+  // const blog = await Blog.findById(...)
+
+  const blog = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)  
+  if (!decodedToken.id) {    
+    return response.status(401).json({ 
+      error: 'token invalid' 
+    })  
+  }  
+
+  if ( blog.user.toString() === decodedToken.id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({
+      error: 'you cannot delete blogs you did not create'
+    })
+  }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
